refactor(ui): move components/document_tree to dbflux_components (S4b) (#102)
