+++ crates/dbflux_components/Cargo.toml
@@ -17,6 +17,7 @@ dbflux_core.workspace = true
 thiserror.workspace = true
 log.workspace = true
 serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
 uuid = { workspace = true }
 chrono.workspace = true
