+++ crates/dbflux_components/src/components/mod.rs
@@ -2,3 +2,4 @@
 // This file grows one line per sub-slice as each component is moved.
 
 pub mod data_table;
+pub mod document_tree;
